Update global architecture to fit lib
